requirements.txt
flask
flask-cors
pillow
requests
python-dotenv
numpy
opencv-python-headless
gunicorn
exifread
